Update layers_tree when project is updated
