+++ g3w-admin/openrouteservice/tests/test_api.py
@@ -108,8 +108,8 @@ from qdjango.apps import QGS_SERVER, get_qgs_project
 from qdjango.models import (ConstraintExpressionRule,
                             ConstraintSubsetStringRule, Layer, Project,
                             SessionTokenFilter, SessionTokenFilterLayer,
-                            SingleLayerConstraint)
-from qdjango.tests.base import CURRENT_PATH, QdjangoTestBase, QgisProject
+                            SingleLayerConstraint, buildLayerTreeNodeObject)
+from qdjango.tests.base import CURRENT_PATH, QdjangoTestBase
 from qgis.core import (Qgis, QgsDataSourceUri, QgsProject, QgsProviderRegistry,
                        QgsVectorLayer)
 from qgis.PyQt.QtCore import QTemporaryDir
@@ -222,8 +222,10 @@ class OpenrouteserviceTest(VCRMixin, QdjangoTestBase):
         Project.objects.filter(
             title='Test openrouteservice project').delete()
 
+        toc = buildLayerTreeNodeObject(project.layerTreeRoot())
+
         cls.qdjango_project = Project(
-            qgis_file=File(open(cls.temp_project_path, 'r')), title='Test openrouteservice project', group=cls.project_group)
+            qgis_file=File(open(cls.temp_project_path, 'r')), title='Test openrouteservice project', group=cls.project_group, layers_tree=toc)
         cls.qdjango_project.save()
 
         for layer_id, layer in project.mapLayers().items():
@@ -469,6 +471,15 @@ class OpenrouteserviceTest(VCRMixin, QdjangoTestBase):
         feature = next(layer.getFeatures())
         self.assertEqual(feature.attribute('range_type'), 'time')
         self.assertEqual(feature.attribute('name'), 'my isochrone')
+
+        # Check layer tree
+        # We need to re-read the Project because of the updated layer_tree
+        self.qdjango_project = Project.objects.get(pk=self.qdjango_project.pk)
+        toc = eval(self.qdjango_project.layers_tree)
+        ids = [l['id'] for l in toc]
+        self.assertIn(layer.id(), ids)
+        self.assertIn(
+            layer.id(), self.qdjango_project.qgis_project.layerTreeRoot().dump())
 
         # Check connections
         connections = get_db_connections(self.qdjango_project.qgis_project)
